DDCNL-7626 Removing home from eligiblity controller
