+++ app/views/errors/session_timeout.scala.html
@@ -38,7 +38,7 @@ sessionExpiredPage = true
 
 @p(HtmlContent(
         button(msg="session.timeout.button", 
-            href = Some(controllers.routes.EligibilityController.home.url), 
+            href = Some(controllers.routes.HowItWorksController.home.url),
             id = "timed-out", 
             attrs = Map("data-journey-click" -> "button - click:Signed out - Marriage Allowance - GOV.UK:start again"))
     ), elmId = Some("continue-to-application"))
