+++ conf/app.routes
@@ -51,7 +51,6 @@ GET         /confirm-change-action              controllers.UpdateRelationshipCo
 GET         /finished-change                    controllers.UpdateRelationshipController.finishUpdate
 POST        /confirm-email                      controllers.UpdateRelationshipController.confirmYourEmailActionUpdate
 
-GET         /how-it-works                       controllers.HowItWorksController.howItWorks
 GET         /eligibility-check-pta              controllers.Default.redirect(to = controllers.routes.EligibilityController.eligibilityCheck.url)
 GET         /lower-earner-pta                   controllers.Default.redirect(to = controllers.routes.EligibilityController.lowerEarnerCheck.url)
 GET         /partners-income-pta                controllers.Default.redirect(to = controllers.routes.EligibilityController.partnersIncomeCheck.url)
@@ -59,7 +58,9 @@ GET         /date-of-birth-check-pta            controllers.Default.redirect(to 
 GET         /do-you-live-in-scotland-pta        controllers.Default.redirect(to = controllers.routes.EligibilityController.doYouLiveInScotland.url)
 GET         /do-you-want-to-apply-pta           controllers.Default.redirect(to = controllers.routes.EligibilityController.doYouWantToApply.url)
 
-GET         /                                   controllers.EligibilityController.home
+GET         /                                   controllers.HowItWorksController.home
+GET         /how-it-works                       controllers.HowItWorksController.howItWorks
+
 GET         /eligibility-check                  controllers.EligibilityController.eligibilityCheck
 POST        /eligibility-check                  controllers.EligibilityController.eligibilityCheckAction
 GET         /lower-earner                       controllers.EligibilityController.lowerEarnerCheck
